pyproject.toml
[project]
name = "mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiolimiter>=1.2.1",
    "beautifulsoup4>=4.14.2",
    "elevenlabs>=2.22.1",
    "fastapi>=0.121.1",
    "gtts>=2.5.4",
    "httpx>=0.28.1",
    "langchain>=1.0.5",
    "langchain-anthropic>=1.0.3",
    "langchain-core>=1.0.4",
    "langchain-groq>=1.0.0",
    "langgraph>=1.0.3",
    "ollama>=0.6.0",
    "pydantic>=2.12.4",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "streamlit>=1.51.0",
    "tenacity>=9.1.2",
    "typing-extensions>=4.15.0",
    "uvicorn>=0.38.0",
]
